+++ CHANGELOG.md
@@ -10,10 +10,11 @@ Planned changes/updates/fixes/features for to be part of one of the next _releas
 
 ## [0.3.0] - 2021-0m-xx
 ### Added
-  * ...
+  * Option to write request response to file
 
 ### Fixed
   * Fixed wrong `config-template.yaml`  
+  * Fixed adding headers if no headers set as parameters
 
 ## [0.2.0] - 2021-04-02
 ### Added